refactor(ai): extract account-info-tool to fix chat.ts max-lines
- Move buildUserInfoSpec, createAccountInfoTool, getMergedTools to
  account-info-tool.ts to satisfy eslint max-lines (300) on chat.ts
