+++ apps/fastify/src/routes/ai/chat.ts
@@ -8,16 +8,11 @@ import {
   smoothStream,
   streamText,
   type ToolSet,
-  tool,
 } from 'ai'
-import { eq } from 'drizzle-orm'
 import type { FastifyPluginAsync } from 'fastify'
-import { z } from 'zod'
-import { getDb } from '../../db/index.js'
-import { users } from '../../db/schema/index.js'
 import { env } from '../../lib/env.js'
 import { ErrorResponseSchema } from '../schemas.js'
-import { createBraveSearchTool } from './brave-search.js'
+import { getMergedTools } from './account-info-tool.js'
 import { getProvider, getResolvedProvider } from './provider.js'
 import { isInsufficientCreditsError } from './upstream-error.js'
 
@@ -84,70 +79,6 @@ async function resolveMessages(rawMessages: unknown[], tools: ToolSet): Promise<
   throw new Error(
     'Invalid request: each message must have parts (UIMessage) or content (CoreMessage)',
   )
-}
-
-const userInfoSpecRoot = 'user-info-1'
-
-function buildUserInfoSpec(user: {
-  name: string | null
-  email: string | null
-  image: string | null
-  username: string | null
-  createdAt: Date
-}) {
-  const joinedAt = new Intl.DateTimeFormat('en-US', {
-    month: 'long',
-    day: 'numeric',
-    year: 'numeric',
-  }).format(user.createdAt)
-  return {
-    root: userInfoSpecRoot,
-    elements: {
-      [userInfoSpecRoot]: {
-        type: 'UserInfo',
-        props: {
-          name: user.name ?? null,
-          email: user.email ?? null,
-          image: user.image ?? null,
-          username: user.username ?? null,
-          joinedAt,
-        },
-        children: [],
-      },
-    },
-  } as const
-}
-
-function createAccountInfoTool(userId: string) {
-  return tool({
-    description:
-      'Returns information about the current authenticated account. Use when the user asks who they are, their account details, when they joined, or similar.',
-    inputSchema: z.object({}),
-    execute: async () => {
-      const db = await getDb()
-      const [user] = await db.select().from(users).where(eq(users.id, userId))
-      if (!user) return 'Account not found.'
-      const spec = buildUserInfoSpec(user)
-      const summaryParts = [`You joined in ${spec.elements[userInfoSpecRoot].props.joinedAt}`]
-      if (user.email) summaryParts.push(`Email: ${user.email}`)
-      if (user.name) summaryParts.push(`Name: ${user.name}`)
-      if (user.username) summaryParts.push(`Username: ${user.username}`)
-      return {
-        __render: 'user-info',
-        spec,
-        summary: summaryParts.join('. '),
-      }
-    },
-  })
-}
-
-function getMergedTools(userId: string, log: import('fastify').FastifyBaseLogger): ToolSet {
-  return {
-    getAccountInfo: createAccountInfoTool(userId),
-    ...(env.BRAVE_SEARCH_API_KEY && {
-      braveSearch: createBraveSearchTool(env.BRAVE_SEARCH_API_KEY, log),
-    }),
-  }
 }
 
 const chatRoute: FastifyPluginAsync = async fastify => {
